api/requirements.txt
anyio==3.6.1
black==22.6.0
click==8.1.3
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pydantic==1.9.1
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.2
